moved Model/account.py to nessie/models/account.py; moved accountRequests.py to nessie/accountRequests.py; refactored main.py to tests/test_accountRequests.py
